File rename refactor on package names and file names
