--- src/main/java/com/uc/budgettracker/repositories/TransactionRepository.java
@@ -1,11 +1,0 @@
-package com.uc.budgettracker.repositories;
-
-import com.uc.budgettracker.dto.Transaction;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TransactionRepository extends CrudRepository<Transaction, Long> {
-
-}
